Add modal and update paths
