+++ src/components/LeftSection/components/ButtonGroup.tsx
@@ -1,0 +1,32 @@
+import { buttonsStyle } from "../../../data/buttonsStyle";
+import type { CSSProperties } from "react";
+import styles from "./ButtonGroup.module.css";
+
+export const ButtonGroup = ({
+  onClickButton,
+}: {
+  onClickButton: (style: CSSProperties) => void;
+}) => {
+  return (
+    <div className={styles["button-group"]}>
+      <h2 className={styles["left-section__subtitle"]}>
+        Choose a <span className={styles["languages__title-span"]}>B</span>utton
+      </h2>
+      <div className={styles["left-section__buttons"]}>
+        {buttonsStyle.map((element) => {
+          return (
+            <div key={element.id} className={styles["button__container"]}>
+              <button
+                className={styles["button__item"]}
+                onClick={() => onClickButton(element.style)}
+                style={element.style}
+              >
+                Click me!
+              </button>
+            </div>
+          );
+        })}
+      </div>
+    </div>
+  );
+};
