+++ src/components/LeftSection/components/Profile.module.css
@@ -1,0 +1,11 @@
+.profile {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.profile__signup {
+  border: none;
+  background-color: var(--color-background-white);
+  border-radius: 0.4rem;
+}
