+++ src/components/LeftSection/components/Profile.tsx
@@ -1,0 +1,10 @@
+import styles from "./Profile.module.css";
+
+export const Profile = () => {
+  return (
+    <div className={styles["profile"]}>
+      <button className={styles["profile__signup"]}>Sign Up</button>
+      <button className={styles["profile__signup"]}>Sign In</button>
+    </div>
+  );
+};
