+++ src/components/WorkBench/WorkBench.module.css
@@ -10,6 +10,7 @@
   flex-direction: column;
   background-color: var(--color-background-black);
   padding: 4rem 2rem 2rem 2rem;
+  justify-content: space-between;
 }
 
 .workbench__right {
